+++ atlasdb-api/src/main/java/com/palantir/atlasdb/keyvalue/api/watch/NoOpInternalLockWatchManager.java
@@ -19,30 +19,26 @@ package com.palantir.atlasdb.keyvalue.api.watch;
 import java.util.Optional;
 import java.util.Set;
 
-import com.palantir.atlasdb.timelock.api.LockWatchRequest;
-import com.palantir.lock.client.NamespacedConjureLockWatchingService;
 import com.palantir.lock.watch.IdentifiedVersion;
-import com.palantir.lock.watch.LockWatchEventCache;
 import com.palantir.lock.watch.LockWatchReferences;
+import com.palantir.lock.watch.NoOpLockWatchEventCache;
 import com.palantir.lock.watch.TransactionsLockWatchEvents;
 
-public final class LockWatchManagerImpl implements LockWatchManager {
-    private final NamespacedConjureLockWatchingService lockWatcher;
-    private final LockWatchEventCache cache;
+public final class NoOpInternalLockWatchManager implements InternalLockWatchManager {
+    public static final InternalLockWatchManager INSTANCE = new NoOpInternalLockWatchManager();
 
-    public LockWatchManagerImpl(NamespacedConjureLockWatchingService lockWatcher, LockWatchEventCache cache) {
-        this.lockWatcher = lockWatcher;
-        this.cache = cache;
+    private NoOpInternalLockWatchManager() {
+        // ...
     }
 
     @Override
-    public void registerWatches(Set<LockWatchReferences.LockWatchReference> lockWatchEntries) {
-        lockWatcher.startWatching(LockWatchRequest.of(lockWatchEntries));
+    public void registerWatches(Set<LockWatchReferences.LockWatchReference> lockWatchReferences) {
+        throw new UnsupportedOperationException("Lock watch registration not supported");
     }
 
     @Override
     public TransactionsLockWatchEvents getEventsForTransactions(Set<Long> startTimestamps,
-            Optional<IdentifiedVersion> version) {
-        return cache.getEventsForTransactions(startTimestamps, version);
+            Optional<IdentifiedVersion> lastKnownVersion) {
+        return NoOpLockWatchEventCache.INSTANCE.getEventsForTransactions(startTimestamps, lastKnownVersion);
     }
 }
