+++ atlasdb-api/src/main/java/com/palantir/atlasdb/transaction/api/TransactionManager.java
@@ -21,7 +21,7 @@ import java.util.function.Supplier;
 import com.palantir.atlasdb.cleaner.api.Cleaner;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.watch.LockWatchManager;
-import com.palantir.atlasdb.keyvalue.api.watch.NoOpLockWatchManager;
+import com.palantir.atlasdb.keyvalue.api.watch.NoOpInternalLockWatchManager;
 import com.palantir.atlasdb.metrics.Timed;
 import com.palantir.atlasdb.transaction.service.TransactionService;
 import com.palantir.exception.NotInitializedException;
@@ -322,7 +322,7 @@ public interface TransactionManager extends AutoCloseable {
 
     @DoNotDelegate
     default LockWatchManager getLockWatchManager() {
-        return NoOpLockWatchManager.INSTANCE;
+        return NoOpInternalLockWatchManager.INSTANCE;
     }
 
     /**
