+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/transaction/impl/SnapshotTransactionManagerTest.java
@@ -44,7 +44,7 @@ import com.palantir.atlasdb.cache.DefaultTimestampCache;
 import com.palantir.atlasdb.cleaner.api.Cleaner;
 import com.palantir.atlasdb.debug.ConflictTracer;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
-import com.palantir.atlasdb.keyvalue.api.watch.NoOpLockWatchManager;
+import com.palantir.atlasdb.keyvalue.api.watch.NoOpInternalLockWatchManager;
 import com.palantir.atlasdb.sweep.queue.MultiTableSweepQueueWriter;
 import com.palantir.atlasdb.transaction.ImmutableTransactionConfig;
 import com.palantir.atlasdb.transaction.api.AtlasDbConstraintCheckingMode;
@@ -80,7 +80,7 @@ public class SnapshotTransactionManagerTest {
             keyValueService,
             new LegacyTimelockService(timestampService, closeableLockService,
                     LockClient.of("lock")),
-            NoOpLockWatchManager.INSTANCE,
+            NoOpInternalLockWatchManager.INSTANCE,
             timestampService,
             closeableLockService,
             mock(TransactionService.class),
@@ -135,7 +135,7 @@ public class SnapshotTransactionManagerTest {
                 keyValueService,
                 new LegacyTimelockService(ts, closeableLockService,
                         LockClient.of("lock")),
-                NoOpLockWatchManager.INSTANCE,
+                NoOpInternalLockWatchManager.INSTANCE,
                 ts,
                 mock(LockService.class), // not closeable
                 mock(TransactionService.class),
@@ -266,7 +266,7 @@ public class SnapshotTransactionManagerTest {
                 metricsManager,
                 keyValueService,
                 timelockService,
-                NoOpLockWatchManager.INSTANCE,
+                NoOpInternalLockWatchManager.INSTANCE,
                 timestampService,
                 mock(LockService.class), // not closeable
                 mock(TransactionService.class),
